requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
DAWG-Python==0.7.2
docopt==0.6.2
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.17.0
fsspec==2025.2.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.3
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyclipper==1.3.0.post6
pymorphy2==0.9.1
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.417127.4579844
PyMuPDF==1.25.2
pyTelegramBotAPI==4.26.0
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
six==1.17.0
sympy==1.13.1
tifffile==2024.8.30
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
